[scanner] refactor: split high-complexity component(s) (#21171)
Split WatchDetailModal.tsx (455 lines) into cohesive sibling files
under web/src/components/stellar/watchDetail/:

- helpers.ts: pure helpers (severityColor, formatRelative,
  formatDuration, deploymentNameFromPodName, matchesWatch, footerBtn)
  and shared constants.
- WatchDetailPrimitives.tsx: small presentational primitives (Tag,
  Stat, SectionHeader, Section, Recommendation) — preserves the
  existing t('stellar.watchDetail.executeViaChat') i18n call.
- WatchDetailHeader.tsx: title + tag row.
- WatchDetailContent.tsx: scrollable body (stats, latest observation,
  recommendations, Stellar's actions, event timeline).
- WatchDetailFooter.tsx: action buttons.

Public API (WatchDetailModal props and exports) is unchanged; all
logic and behavior are preserved. Mirrors the sibling-directory
pattern used in #21180 (WhatsNewModal) and #21182 (EventCard).
